refactor:  add calculate d-day util
